+++ api/src/index.js
@@ -59,6 +59,7 @@ app.use("/person", require("./controllers/person"));
 app.use("/action", require("./controllers/action"));
 app.use("/structure", require("./controllers/structure"));
 app.use("/comment", require("./controllers/comment"));
+app.use("/passage", require("./controllers/passage"));
 app.use("/report", require("./controllers/report"));
 app.use("/place", require("./controllers/place"));
 app.use("/relPersonPlace", require("./controllers/relPersonPlace"));