+++ api/src/models/organisation.js
@@ -18,6 +18,9 @@ Organisation.init(
     customFieldsObs: DataTypes.JSONB,
     customFieldsPersonsSocial: DataTypes.JSONB,
     customFieldsPersonsMedical: DataTypes.JSONB,
+    migrating: { type: DataTypes.BOOLEAN, default: false },
+    migrations: DataTypes.ARRAY(DataTypes.TEXT),
+    migrationLastUpdateAt: DataTypes.DATE,
   },
   { sequelize, modelName: "Organisation", freezeTableName: true }
 );